bugfix: handle being asked to open a directory
It was crashing, now it just returns an empty buffer.

	modified:   src/fileaccessor.rb

+++ src/fileaccessor.rb
@@ -40,7 +40,7 @@ class FileAccessor
 	# Read file into an array.
 	def read
 
-		return([""]) if @name=="" || !File.exists?(@name)
+		return([""]) if @name=="" || !File.exists?(@name) || File.directory?(@name)
 
 		begin
 			text = File.open(@name,"rb:UTF-8"){|f| f.read}
